refactor: 잘못된 캐스팅 제거 및 서비스 위치 변경
