+++ src/main/java/com/example/barrier_free/domain/favorite/controller/FavoriteController.java
@@ -3,11 +3,14 @@ package com.example.barrier_free.domain.favorite.controller;
 import java.util.List;
 
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.barrier_free.domain.favorite.FavoriteService;
+import com.example.barrier_free.domain.favorite.dto.FavoritePlaceGroupResponse;
 import com.example.barrier_free.domain.favorite.dto.FavoriteRequestDto;
 import com.example.barrier_free.domain.favorite.dto.PlaceResponse;
 import com.example.barrier_free.global.response.ApiResponse;
@@ -47,6 +50,15 @@ public class FavoriteController {
 	public ApiResponse<?> getWeeklyTop3() {
 		List<PlaceResponse> weeklyTop3 = favoriteService.getWeeklyTop3();
 		return ApiResponse.success(SuccessCode.OK, weeklyTop3);
+	}
+
+	@GetMapping(value = "/users/{userId}/favorites")
+	public ApiResponse<?> getFavorite(@PathVariable Long userId,
+		@RequestParam(required = false) List<Integer> facilities
+	) {
+		FavoritePlaceGroupResponse response = favoriteService.getFavorite(userId, facilities);
+		return ApiResponse.success(SuccessCode.OK, response);
+
 	}
 
 }
